rename encryptionkey resource to contextspec
