+++ service/controller/ignition_resource_set.go
@@ -14,7 +14,7 @@ import (
 	"github.com/giantswarm/ignition-operator/pkg/project"
 	"github.com/giantswarm/ignition-operator/service/controller/controllercontext"
 	"github.com/giantswarm/ignition-operator/service/controller/key"
-	"github.com/giantswarm/ignition-operator/service/controller/resource/encryptionkey"
+	"github.com/giantswarm/ignition-operator/service/controller/resource/contextspec"
 	"github.com/giantswarm/ignition-operator/service/controller/resource/templatefiles"
 	"github.com/giantswarm/ignition-operator/service/controller/resource/templateignition"
 	"github.com/giantswarm/ignition-operator/service/controller/resource/templateunits"
@@ -30,12 +30,12 @@ func newIgnitionResourceSet(config ignitionResourceSetConfig) (*controller.Resou
 
 	var encryptionkeyResource resource.Interface
 	{
-		c := encryptionkey.Config{
+		c := contextspec.Config{
 			K8sClient: config.K8sClient,
 			Logger:    config.Logger,
 		}
 
-		encryptionkeyResource, err = encryptionkey.New(c)
+		encryptionkeyResource, err = contextspec.New(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
