+++ service/controller/resource/contextspec/create.go
@@ -1,4 +1,4 @@
-package encryptionkey
+package contextspec
 
 import (
 	"context"
@@ -98,7 +98,6 @@ func crToCCSpec(cr v1alpha1.Ignition) controllercontext.ContextSpec {
 			},
 			Etcd: controllercontext.ContextSpecEtcd{
 				Domain: cr.Spec.Etcd.Domain,
-				Image:  cr.Spec.Etcd.Image,
 				Port:   cr.Spec.Etcd.Port,
 				Prefix: cr.Spec.Etcd.Prefix,
 			},
@@ -122,7 +121,6 @@ func crToCCSpec(cr v1alpha1.Ignition) controllercontext.ContextSpec {
 				Kubelet: controllercontext.ContextSpecKubernetesKubelet{
 					Domain: cr.Spec.Kubernetes.Kubelet.Domain,
 				},
-				Image:   cr.Spec.Kubernetes.Image,
 				IPRange: cr.Spec.Kubernetes.IPRange,
 			},
 			Provider: cr.Spec.Provider,
