+++ service/controller/resource/contextspec/delete.go
@@ -1,4 +1,4 @@
-package encryptionkey
+package contextspec
 
 import (
 	"context"
